monitoring_service/pyproject.toml
[tool.poetry]
name = "monitoring_service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
gino = {extras = ["starlette"], version = "^1.0.1"}
fastapi = "^0.61.0"
uvicorn = "^0.11.8"
gunicorn = "^20.0.4"
alembic = "^1.4.2"
psycopg2 = "^2.8.5"
celery = "^4.4.7"
celery_pool_asyncio = "^0.2.0"
python-dotenv = "^0.14.0"
isort = "^5.4.2"
flake8 = "^3.8.3"
black = "^19.10b0"

[tool.poetry.dev-dependencies]

[tool.poetry.plugins."monitoring_service.modules"]
"events" = "monitoring_service.views.events"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

